[update]增加version页面分页功能

+++ src/pages/version/components/versionContent.jsx
@@ -2,7 +2,16 @@ import React, { Component } from 'react';
 import { withRouter, Link } from 'react-router-dom';
 import store from '../../../mobx/versionStore';
 import PropTypes from 'prop-types';
-import { Descriptions, Badge, Modal, Divider, Button, Tag, Icon } from 'antd';
+import {
+  Descriptions,
+  Badge,
+  Modal,
+  Divider,
+  Button,
+  Tag,
+  Icon,
+  Empty
+} from 'antd';
 import { observer } from 'mobx-react';
 const { confirm } = Modal;
 @observer
@@ -38,38 +47,42 @@ class Content extends Component {
   };
   render() {
     const list = store.list;
-    return list.map((version, index) => {
-      return (
-        <Descriptions
-          title={<Tag color="geekblue">{version.versionName}</Tag>}
-          layout="vertical"
-          bordered
-          key={index}
-        >
-          <Descriptions.Item label="版本描述" className="version-desc">
-            {version.versionDesc}
-          </Descriptions.Item>
-          <Descriptions.Item label="版本状态" className="version-status">
-            <Badge
-              status={index === 0 ? 'processing' : 'default'}
-              text={index === 0 ? '新版本' : '旧版本'}
-            />
-          </Descriptions.Item>
-          <Descriptions.Item className="version-atcions" label="版本操作">
-            <Link to={`/product/project/version/${version.versionId}/error`}>
-              <Button type="primary">
-                <Icon type="line-chart" /> 查看
+    if (list.length === 0) {
+      return <Empty imageStyle={{ height: '25rem' }} />;
+    } else {
+      return list.map((version, index) => {
+        return (
+          <Descriptions
+            title={<Tag color="geekblue">{version.versionName}</Tag>}
+            layout="vertical"
+            bordered
+            key={index}
+          >
+            <Descriptions.Item label="版本描述" className="version-desc">
+              {version.versionDesc}
+            </Descriptions.Item>
+            <Descriptions.Item label="版本状态" className="version-status">
+              <Badge
+                status={index === 0 ? 'processing' : 'default'}
+                text={index === 0 ? '新版本' : '旧版本'}
+              />
+            </Descriptions.Item>
+            <Descriptions.Item className="version-atcions" label="版本操作">
+              <Link to={`/product/project/version/${version.versionId}/error`}>
+                <Button type="primary">
+                  <Icon type="line-chart" /> 查看
+                </Button>
+              </Link>
+              <Divider type="vertical" />
+              <Button type="danger">
+                <Icon type="delete" />
+                删除
               </Button>
-            </Link>
-            <Divider type="vertical" />
-            <Button type="danger">
-              <Icon type="delete" />
-              删除
-            </Button>
-          </Descriptions.Item>
-        </Descriptions>
-      );
-    });
+            </Descriptions.Item>
+          </Descriptions>
+        );
+      });
+    }
   }
 }
 export default withRouter(Content);
